relocate test files
