+++ packages/sdk/src/allowances.test.ts
@@ -1,9 +1,10 @@
-import { AbiCoder, concat, hexlify, zeroPadValue } from "ethers"
 import { expect, it, suite } from "vitest"
+import { AbiCoder, concat, hexlify, zeroPadValue } from "ethers"
 import { Allowance } from "zodiac-roles-deployments"
 
+import { applyAllowances } from "./allowances"
+
 import { Roles__factory } from "../../evm/typechain-types"
-import { applyAllowances } from "../src/allowances"
 
 const iface = Roles__factory.createInterface()
