+++ packages/sdk/src/calls/removeObsoleteCalls.test.ts
@@ -1,8 +1,9 @@
 import { expect, it, suite } from "vitest"
 import { ExecutionOptions } from "zodiac-roles-deployments"
 
-import { removeObsoleteCalls } from "../src/calls"
-import { Call } from "../src/calls/types"
+import { removeObsoleteCalls } from "./removeObsoleteCalls"
+
+import { Call } from "./types"
 
 suite("removeObsoleteCalls", () => {
   it("should remove function permission updates if later on the entire target is cleared", () => {
