+++ packages/sdk/src/conditions/conditionId.test.ts
@@ -1,7 +1,8 @@
 import { expect, it, suite } from "vitest"
 import { Operator, ParameterType } from "zodiac-roles-deployments"
 
-import { conditionAddress, normalizeCondition } from "../src/conditions"
+import { conditionAddress } from "./conditionId"
+import { normalizeCondition } from "./normalizeCondition"
 
 suite("conditionAddress", () => {
   it("calculates the create2 storage address of the condition", () => {
