+++ packages/sdk/src/conditions/normalizeCondition.test.ts
@@ -1,12 +1,11 @@
 import { expect, it, suite } from "vitest"
 import { Operator, ParameterType } from "zodiac-roles-deployments"
 
-import { normalizeCondition } from "../src/conditions"
-import { stripIds } from "../src/conditions/normalizeCondition"
-import { FunctionPermissionCoerced, c } from "../src/permissions"
-import { allow } from "../src/entrypoints/kit"
-import { mergeFunctionPermissions } from "../src/permissions/mergeFunctionPermissions"
-import { encodeAbiParameters } from "../src/utils/encodeAbiParameters"
+import { normalizeCondition, stripIds } from "./normalizeCondition"
+import { c, FunctionPermissionCoerced } from "../permissions"
+import { allow } from "../entrypoints/kit"
+import { mergeFunctionPermissions } from "../permissions/mergeFunctionPermissions"
+import { encodeAbiParameters } from "../utils/encodeAbiParameters"
 
 const DUMMY_COMP = (id: number) => ({
   paramType: ParameterType.Static,
